Allow recursive definitions.
This pull request is a little magical, and allows the use of
recursive definitions.

Typically when we create a new word/definition we essentially build
up some "bytecode" which might allow several things:

* The lookup and reference of an existing word being stored.
* Some special pseudo operations, prefixed with negative values.
  * Jumps
  * Inline integers

Compiling a word which contains references to unknown words fails,
because their "address", or "offset" cannot be found.  Which means
this fails:

        : factorial dup 1 >  if  dup 1 -  factorial *  then  ;

The compilation will encode "dupe", "1", ">", etc, but will then
fail to find the word "factorial" towards the end - that word doesn't
exist, because the definition is not complete until we hit ";".

This pull-request makes an assumption; if a definition is _recursive_
then the call to any unknown word will be a call to "itself", and thus
will be the final word in our known set of words.

i.e. When compiling a recursive word we calculate the word's offset
which will be compiled, and use it.  This allow:

        : factorial recursive  dup 1 >  if  dup 1 -  factorial *  then  ;

And this closes #16.